chore: rename JobSchedular to JobScheduler
